requirements.txt
torch>=2.1.0
Pillow>=9.0.0
torchvision>=0.16.0
sentencepiece>=0.1.99
